requirements.txt
aiohttp==3.2.1
alabaster==0.7.9
altgraph==0.15
anaconda-client==1.6.0
anaconda-project==0.4.1
anytree==2.4.3
appdirs==1.4.3
astroid==1.4.9
astropy==2.0
async-timeout==3.0.1
atari-py==0.1.7
attrs==18.2.0
auto-py-to-exe==1.3.0
Babel==2.3.4
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.6.0
bitarray==0.8.1
bleach==1.5.0
boto==2.45.0
bottle==0.12.13
bottle-websocket==0.2.9
Bottleneck==1.2.1
bz2file==0.98
cairocffi==0.9.0
CairoSVG==2.1.3
cffi==1.11.5
chardet==2.3.0
chest==0.2.3
click==6.7
cloudpickle==0.2.2
clyent==1.2.2
cmake==3.13.3
colorama==0.3.7
comtypes==1.1.2
conda==4.4.6
configobj==5.0.6
contextlib2==0.5.4
cryptography==1.7.1
cssselect2==0.2.1
cycler==0.10.0
Cython==0.28.3
cytoolz==0.8.2
dask==1.0.0
datashape==0.5.4
decorator==4.4.0
defusedxml==0.5.0
dill==0.2.5
distributed==1.25.0
dlib==19.17.0
docutils==0.13.1
Eel==0.9.10
emoji==0.5.0
enum34==1.1.6
et-xmlfile==1.0.1
fastcache==1.0.2
Flask==0.12
Flask-Cors==3.0.2
future==0.16.0
gensim==3.2.0
gevent==1.2.1
gevent-websocket==0.10.1
glfw==1.7.0
gmusicapi==11.0.0
gpsoauth==0.4.1
graphviz==0.8.3
greenlet==0.4.11
gym==0.10.9
h5py==2.7.1
HeapDict==1.0.0
html5lib==0.9999999
httplib2==0.10.3
idna==2.2
idna-ssl==1.1.0
imageio==2.5.0
imagesize==0.7.1
imbalanced-learn==0.3.1
ipykernel==4.5.2
ipyparallel==6.0.2
ipython==6.1.0
ipython-genutils==0.1.0
ipywidgets==5.2.2
isort==4.2.5
itsdangerous==0.24
jdcal==1.3
jedi==0.10.2
Jinja2==2.9.4
joblib==0.11
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==4.4.0
jupyter-console==5.0.0
jupyter-core==4.2.1
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
lazy-object-proxy==1.2.2
llvmlite==0.18.0
locket==0.2.0
lockfile==0.12.2
lxml==3.7.2
macholib==1.9
Markdown==2.6.9
MarkupSafe==0.23
matplotlib==2.2.2
MechanicalSoup==0.9.0.post4
menuinst==1.4.4
mistune==0.7.3
mock==2.0.0
mosspy==1.0.5
mpmath==0.19
msgpack-python==0.4.8
mujoco-py==1.50.1.68
multidict==4.2.0
multipledispatch==0.4.9
mutagen==1.39
nbconvert==4.2.0
nbformat==4.2.0
networkx==2.2
nltk==3.2.2
nose==1.3.7
notebook==5.0.0
numba==0.33.0
numexpr==2.6.2
numpy==1.12.1
numpydoc==0.6.0
oauth2client==4.1.2
odo==0.5.0
olefile==0.44
opencv-python==4.0.0.21
openpyxl==2.4.1
packaging==18.0
pandas==0.23.3
partd==0.3.8
path.py==0.0.0
pathlib2==2.2.0
patsy==0.4.1
pbr==3.1.1
pefile==2017.11.5
pep8==1.7.0
pickleshare==0.7.4
Pillow==5.4.0
playsound==1.2.2
ply==3.9
prettytable==0.7.2
proboscis==1.2.6.0
progressbar2==3.37.1
prompt-toolkit==1.0.9
protobuf==3.5.1
psutil==5.0.1
py==1.4.32
py2exe==0.9.2.2
pyasn1==0.1.9
pyasn1-modules==0.2.1
pycosat==0.6.3
pycparser==2.17
pycrypto==2.6.1
pycryptodomex==3.4.7
pycurl==7.43.0
pyemd==0.4.3
pyflakes==1.5.0
pygame==1.9.3
pyglet==1.2.4
Pygments==2.1.3
PyInstaller==3.3.1
pylint==1.6.4
PyOpenGL==3.1.0
pyOpenSSL==16.2.0
pyparsing==2.1.4
pytest==3.0.5
python-chess==0.23.5
python-dateutil==2.6.0
python-Levenshtein==0.12.0
python-resize-image==1.1.11
python-utils==2.3.0
pytz==2016.10
PyWavelets==0.5.2
pywin32==220
PyYAML==3.12
pyzmq==16.0.2
QtAwesome==0.4.3
qtconsole==4.2.1
QtPy==1.2.1
requests==2.12.4
rope-py3k==0.9.4.post1
rsa==3.4.2
scikit-image==0.15.0
scikit-learn==0.19.0
scipy==0.19.1
seaborn==0.7.1
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
smart-open==1.5.3
snowballstemmer==1.2.1
sockjs-tornado==1.0.3
somoclu==1.7.5
sortedcontainers==1.5.7
sphinx==1.5.1
spyder==3.1.2
SQLAlchemy==1.1.5
statsmodels==0.8.0
sympy==1.0
tables==3.4.2
tblib==1.3.2
tensorflow==1.4.0
tensorflow-gpu==1.1.0
tensorflow-tensorboard==0.4.0rc3
tflearn==0.3.2
tinycss2==0.6.1
toolz==0.8.2
tornado==4.5.2
traitlets==4.3.1
unicodecsv==0.14.1
UNKNOWN==0.0.0
validictory==1.1.2
virtualenv==16.4.3
wcwidth==0.1.7
webencodings==0.5
Werkzeug==0.11.15
whichcraft==0.4.1
widgetsnbextension==1.2.6
win-unicode-console==0.5
wordcloud==1.4.1
wrapt==1.10.8
xlrd==1.0.0
XlsxWriter==0.9.6
xlwings==0.10.2
xlwt==1.2.0
yarl==1.2.6
zict==0.1.3
